requirements.txt
certifi==2021.5.30
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
PyAudio==0.2.11
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.2
pyrtlsdr==0.2.92
pysdr==1.2
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
